paperclip/ validações / tradução / mensagem de sucesso
- Adição da gem do paperclip
- Criação da migration para adicionar as colunas do paperclip ao banco
- Implementação das validações do paperclip
- Ajustes na validações de pagamento
- Adição da mensagem de sucesso na pagina inicial
- Configuração para internacionalização da aplicação para português
- Adição do arquivo pt-BR.yml
- Alteração das inflections para português

+++ db/migrate/20151205165428_create_candidatos.rb
@@ -9,8 +9,8 @@ class CreateCandidatos < ActiveRecord::Migration
       t.string :camisa
       t.string :contato_emergencia
       t.string :email
-      t.integer :telefone
-      t.integer :telefone_contato_emergencia
+      t.decimal :telefone, precision: 11
+      t.decimal :telefone_contato_emergencia, precision: 11
       t.boolean :pagamento_efetuado
       t.timestamps null: false
     end
